requirements.txt
yfinance
pandas
numpy
matplotlib>=3.8
python-dateutil
requests
beautifulsoup4
python-dotenv
gspread
google-api-python-client
google-auth
